Add room database instance and DAO interface

+++ Mobile/app/src/main/java/com/csis4495/aiorestaurants/db/MoshiEntities.kt
@@ -22,7 +22,7 @@ data class RestaurantJson(
 )
 
 data class UserCategoryJson(
-    var categoryId: Double,
+    var categoryId: Int,
     var category: String
 )
